finished work...

+++ pages/api/categories/[id].js
@@ -49,6 +49,12 @@ const deleteCategory = async (req, res) => {
 
     const { id } = req.query;
 
+    const products = await Products.findOne({ category: id });
+    if (products)
+      return res.status(400).json({
+        err: "Please delete all products with a relationship",
+      });
+
     await Categories.findByIdAndDelete(id);
 
     res.json({ msg: "Success! Deleted a category" });
